refactor(insights): break baseline-store↔statistical-baseline import cycle
Plan 48 introduced a circular dependency (baseline-store imports the Baseline
type from statistical-baseline, which imports the store's getBaseline/upsertBaseline
for refit) — a no-circular depcruise violation that landed because depcruise is
non-blocking for auto-merge. Extract the shared types (Baseline, BaselineConfidence,
AnomalyScore) into a leaf baseline-types module; re-export from statistical-baseline
for existing consumers. No behavior change.
