+++ apps/dashboard/src/lib/insights/statistical-baseline.ts
@@ -13,34 +13,23 @@
  * static threshold. See plans/48-statistical-anomaly-baselines.md.
  */
 
+import type {
+  AnomalyScore,
+  Baseline,
+  BaselineConfidence,
+} from './baseline-types'
+
 import { getBaseline, upsertBaseline } from './baseline-store'
 
-/** A fitted per-host/per-metric baseline distribution. */
-export interface Baseline {
-  readonly hostId: string
-  readonly metric: string
-  readonly mean: number
-  readonly stddev: number
-  readonly median: number
-  readonly mad: number
-  readonly sampleCount: number
-  readonly windowStart: number
-  readonly fittedAt: number
-}
-
-/** Confidence in a baseline, driven by how many samples it was fit from. */
-export type BaselineConfidence = 'low' | 'medium' | 'high'
-
-/** Result of scoring a value against a baseline. */
-export interface AnomalyScore {
-  /** Standard deviations from the baseline mean; `0` when no baseline was used. */
-  readonly z: number
-  /** `true` when `|z|` exceeds the threshold (default 2). */
-  readonly isAnomaly: boolean
-  readonly confidence: BaselineConfidence
-  /** `false` means the caller must fall back to its static threshold (cold start / degenerate baseline). */
-  readonly usedBaseline: boolean
-}
+// The Baseline/AnomalyScore/BaselineConfidence types live in the leaf module
+// `baseline-types` so the store and this fitter can share them without importing
+// each other (avoids a no-circular depcruise violation). Re-exported here so
+// existing consumers can keep importing them from `statistical-baseline`.
+export type {
+  AnomalyScore,
+  Baseline,
+  BaselineConfidence,
+} from './baseline-types'
 
 /** Modified z-score cutoff (Iglewicz & Hoaglin) for rejecting outliers before fitting — the IQR-fence equivalent. */
 const ROBUST_Z_OUTLIER_CUTOFF = 3.5
